Focus search box in onResume().
This fixes a problem where the search box is not focused when QSB
is first started. Since the text box is the only place you can type,
it seems reasonable that it is focused on every entry to the app,
even if returning using BACK and something else (button, suggestion)
was focused before.