Add a suggestions footer view
Also, refactor source and corpus creation. This turned out not
to be needed to fix the bug, but I think it was a good thing anyway.

Also, include a suggestion type in the qsb_click log.
This is only set for Genie suggestions at the moment.

Fixes http://b/issue?id=2445622

Change-Id: I332c0566fedecccaae194304310c1f13d7201e2d