Drop "query jamming" feature from QSB
Before, QSB put the text of web suggestions in the query field when
selecting a suggestion with the DPAD or trackball. The code for this
was complex and error-prone (see for example http://b/issue?id=2515755
and http://b/issue?id=2748984). Now there is an on-screen button to
put the suggestion text in the query field. Thus we should remove this
feature.

Bug http://b/issue?id=2776271

Change-Id: Iebafc24cffca0c7cf113e426564ad22bb93da738