More edit distance optimization
Reduces CPU time for formatSuggestion() in LevenshteinFormatterTest from 44 ms
to 30 ms (32% reduction). This is on top of mathewi's recent optimization. The
total reduction is now from 72 ms to 30 ms (58% reduction).

Also adds @SmallTest annotations to the edit distance tests so that
they will be included in the continuous test suite.

Bug: 2836498
Change-Id: I7cc4156df4f9e993ce72cd35001d1abaa7c239e3