Only refresh shortcuts that are displayed in the UI.
Change I6413d82f9051bb64487c65817bfbcdea0f0c3097 introduced a problem whereby
*all* the shortcuts were updated, as the shortcut repository did not filter them,
and it refreshed all the shortcuts it returned. Now, the refreshing is started
after the suggestion promotion stage, so that only shortcuts that are actually
displayed in the suggestions list will be refreshed. This results in far fewer
refreshes and database updates (which can be time consuming).

Change-Id: I4b4cac522d12a4e304ddebfca7466f670cc3811b