Load icons from other processes in a seperate thread.
This protects againts the process not responding, which previously caused QSB to
lock up. We also protect against the process throwing an exception in the same
case.

This change causes some icons to load after the suggestions are displayed. For
example, contact photos are displayed shortly after the contacts show, but the
quick contact badge is visible immediately. This only happens for suggestions
with content:// URIs for the icons.

Bug: 3064103

Change-Id: I23f36a82d381ea76897d65fe2f3ec46f6642b7d5