Clean up suggestion layouts and fix spacing
This adds a special layout for web suggestions,
with fewer views, and moves all the web search suggestion specific
code out of DefaultSuggestionView.
It also gets rid of the per-source SuggestionViewFactories and creates
a global factory that delegates to other factories. This is to make
the choice of view dependent on the suggestion type instead of the
source.

Bug: 3202129
Bug: 3202115
Change-Id: Ie1963acd90bedda523f6c25c26b76b6a075b41af